refactor: add common/domain/util
