+++ androidApp/src/main/java/com/example/tasky/android/agenda/viewmodel/AgendaDetailsViewModel.kt
@@ -19,8 +19,8 @@ import com.example.tasky.android.agenda.screen.AgendaDetailsScreenEvent
 import com.example.tasky.android.agenda.screen.AgendaDetailsScreenState
 import com.example.tasky.android.agenda.screen.AgendaDetailsScreenType
 import com.example.tasky.android.utils.IImageCompressor
+import com.example.tasky.common.domain.util.toLocalDateTime
 import com.example.tasky.common.model.onSuccess
-import com.example.tasky.common.util.toLocalDateTime
 import com.example.tasky.login.domain.manager.SessionManager
 import kotlinx.coroutines.channels.Channel
 import kotlinx.coroutines.flow.MutableStateFlow
